eda/data-pipeline/requirements.txt
# OT data_pipeline requirements
addict==2.2.0
attrs==19.1.0
biopython==1.73
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
codecov==2.0.15
ConfigArgParse==0.14.0
coverage==4.5.3
decorator==4.4.0
dill==0.2.5
elasticsearch-dsl==7.0.0
#this is a backport and should be removed for py3
#this syntax only works for pip > 6
functools32==3.2.3.post2 ; python_version < '3'
future==0.17.1
idna==2.8
isodate==0.6.0
jsonpickle==1.1
jsonschema==3.0.0a3
lxml==4.3.2
more-itertools==5.0.0
networkx==2.2
numpy==1.16.2
petl==1.2.0
psutil==5.6.1
pyparsing==2.3.1
pypeln==0.1.6
pyrsistent==0.14.11
python-dateutil==2.8.0
PyYAML==5.1
rdflib==4.2.2
requests==2.21.0
requests-file==1.4.3
rfc3987==1.3.8
scikit-learn==0.20.3
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
tabulate==0.8.3
urllib3==1.24.1
Yapsy==1.12.0
opentargets-ontologyutils==1.2.2
opentargets-urlzsource==1.0.0
opentargets-validator==0.4.0

# Deps for visualization and exploration
jupyterlab
matplotlib
pandas
plotnine
seaborn
